Move environment from Script to Oneshot/Longrun
Instead of storing a reference to the same Enviroment in the
Script objects, store it in the Oneshot/Longrun class. This
simplify ScriptBuilder, LoggerScriptBuilder, Script and
LoggerScriptBuilder classes, without any major drawback. The only
exception is that now when generating script code, it might be needed
to pass both Script and Enviroment.